requirements.txt
nlpcloud
numpy
python-dotenv
spyder